refactor(examples): keep shaders in app entry points
