+++ examples/tutorials/shader-plugins/app.ts
@@ -19,6 +19,8 @@ import {
   fillPatternShaderPlugin
 } from '../../fill-pattern-shader-plugin';
 
+const {source, vs, fs} = getShaderSources();
+
 type PatternTriangle = {
   center: [number, number];
   fillPattern: FillPatternType;
@@ -60,8 +62,6 @@ const TRIANGLE_VERTICES = [
   {position: [0.24, -0.2] as [number, number], uv: [1, 0] as [number, number]},
   {position: [0, 0.26] as [number, number], uv: [0.5, 1] as [number, number]}
 ];
-
-import {fs, source, vs} from './shaders';
 
 export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
   static info = makeExamplePanelHostHtml();
@@ -270,4 +270,95 @@ function makeClipSettingsSchema(): SettingsSchema {
       }
     ]
   };
+}
+
+function getShaderSources() {
+  const source = /* wgsl */ `\
+  struct VertexOutput {
+    @builtin(position) position: vec4<f32>,
+    @location(0) fillPatternType: f32,
+    @location(1) fillPatternSize: vec2<f32>,
+    @location(2) fillPatternUv: vec2<f32>,
+  };
+
+  @vertex
+  fn vertexMain(
+    @location(0) position: vec2<f32>,
+    @location(1) fillPatternType: f32,
+    @location(2) fillPatternSize: vec2<f32>,
+    @location(3) fillPatternUv: vec2<f32>,
+    @location(4) triangleCenter: vec2<f32>
+  ) -> VertexOutput {
+    var output: VertexOutput;
+    output.position = vec4<f32>(position, 0.0, 1.0);
+    CLIP_POSITION(&output.position, triangleCenter, position);
+    output.fillPatternType = fillPatternType;
+    output.fillPatternSize = fillPatternSize;
+    output.fillPatternUv = fillPatternUv;
+    return output;
+  }
+
+  @fragment
+  fn fragmentMain(inputs: VertexOutput) -> @location(0) vec4<f32> {
+    let fillColor = vec4<f32>(0.08, 0.09, 0.11, 1.0);
+    var color = pluginApplyFillPattern(
+      fillColor,
+      inputs.fillPatternType,
+      inputs.fillPatternUv,
+      inputs.fillPatternSize
+    );
+    CLIP_COLOR(&color);
+    return color;
+  }
+  `;
+
+  const vs = /* glsl */ `\
+  #version 300 es
+
+  in vec2 position;
+  in float fillPatternType;
+  in vec2 fillPatternSize;
+  in vec2 fillPatternUv;
+  in vec2 triangleCenter;
+
+  out float vFillPatternType;
+  out vec2 vFillPatternSize;
+  out vec2 vFillPatternUv;
+
+  void main() {
+    vFillPatternType = fillPatternType;
+    vFillPatternSize = fillPatternSize;
+    vFillPatternUv = fillPatternUv;
+    gl_Position = vec4(position, 0.0, 1.0);
+    CLIP_POSITION(gl_Position, triangleCenter, position);
+  }
+  `;
+
+  const fs = /* glsl */ `\
+  #version 300 es
+  precision highp float;
+
+  in float vFillPatternType;
+  in vec2 vFillPatternSize;
+  in vec2 vFillPatternUv;
+
+  out vec4 fragColor;
+
+  void main() {
+    vec4 fillColor = vec4(0.08, 0.09, 0.11, 1.0);
+    fragColor = plugin_applyFillPattern(
+      fillColor,
+      vFillPatternType,
+      vFillPatternUv,
+      vFillPatternSize
+    );
+    CLIP_COLOR(fragColor);
+  }
+  `;
+
+  return {
+    source,
+    vs,
+    fs
+  } as const;
 }
